refactor(backend): extract the three leaf projections from OpenBotDatabase
Moves `ProviderSessions`, `ThreadSummaries` and `MailboxProjection` into
`src/backend/database/`. None of the three touches another controller, so they
land in one commit.

`ProviderSession` and `StoredThreadSummary` move with their owners and are
re-exported from the facade; the mailbox interfaces become exported on the
controller because `persistConversationAndMailbox` and `replaceMailboxState`
name `MailboxProjectionState` across the module boundary. That widens the
controller's surface, not the facade's.

The dead `_rebaseHistory` parameter is kept and now forwarded through the
facade delegation rather than dropped: removing it would touch
`mailbox-store.ts`, which this change otherwise leaves closed.
